Revert "[Port] Crew monitoring console sorting (#11363)"
This reverts commit 7429631f4bab55e54e24e055bf3b960549462378.
