+++ tgui/packages/tgui/constants.js
@@ -20,7 +20,6 @@ export const COLORS = {
     science: '#9b59b6',
     engineering: '#f1c40f',
     cargo: '#f39c12',
-    service: '#7cc46a',
     centcom: '#00c100',
     other: '#c38312',
   },
